grok/grok/pipeline/requirements.txt
fabric==1.9.0
PyYAML==3.10